Changes in novel agrD searching
Made novel agrD hmm search a separate step. 
The hmmer search will now be performed if either the agr-group is unable to be determined or if less than 5 kmers match.
The frameshift search will happen even if less than 5 kmers match using the appropriate reference. 
The programme will exit if the agr group type is undeterminable but agr operon extraction will be attempted.